requirements.txt
openai
python-dotenv
tenacity
argparse
python-docx
pydub
nltk
tiktoken
PyPDF2
dotenv